pyproject.toml
[tool.poetry]
name = "social-network"
version = "0.1.0"
description = ""
authors = ["llyndev <weslley.alves3310@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9 || ^3.11"
fastapi = {extras = ["standard"], version = "^0.115.3"}
tortoise-orm = "^0.21.7"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"

[tool.taskipy.tasks]
dev = "fastapi dev dev.py"
prod = "uvicorn src.api.application:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
